Move Parcelable Device to a top level class
